Ads Management System
This is a node JS project developed in Loopback 3 framework with MySQL database. This project also consists of configuration of storage and deployment. The prime purpose of this project is to maintain the counts of promotion Ads and to show them depending on the percentage. Which Ad you want to you on which Web and how many times that specific Ad will show overall or on specific web.